Minor refactor, changed the return type of GetTotalPrice to be a decimal, since it better represents a price than an integer

+++ Checkout/Checkout.cs
@@ -2,7 +2,7 @@
 
 public class Checkout
 {
-    public int GetTotalPrice()
+    public decimal GetTotalPrice()
     {
         return 0;
     }
